requirements.txt
PyYAML>=6.0
requests>=2.28.0
PyQt6>=6.0.0
PyQt6-Charts>=6.0.0
